python-dist/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qntm"
version = "0.2.0"
description = "qntm secure messaging protocol client library and CLI"
readme = "README.md"
license = "BUSL-1.1"
requires-python = ">=3.10"
dependencies = [
    "PyNaCl>=1.5.0",
    "cryptography>=42.0",
    "cbor2>=5.6",
]
authors = [
    { name = "Corpo LLC", email = "hello@corpo.llc" },
]
keywords = ["messaging", "encryption", "security", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Communications",
    "Topic :: Security :: Cryptography",
]

[project.urls]
Homepage = "https://qntm.corpo.llc"
Repository = "https://github.com/corpollc/qntm"

[project.scripts]
qntm = "qntm:main"

[tool.hatch.build.targets.wheel]
packages = ["src/qntm"]

[tool.hatch.build.targets.wheel.shared-data]
